pyproject.toml
[project]
name = "kalipso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "pdfminer-six>=20250506",
    "pip>=25.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]
